+++ tests/test_unknown.py
@@ -1,7 +1,7 @@
 import pytest
 
 from blebox_uniapi.session import ApiHost
-from blebox_uniapi.products import Products
+from blebox_uniapi.box import Box
 from blebox_uniapi import error
 
 from .conftest import json_get_expect
@@ -31,7 +31,7 @@ class TestUnknownDevice:
             port = 80
             timeout = 2
             api_host = ApiHost(host, port, timeout, aioclient_mock, None, None)
-            await Products.async_from_host(api_host)
+            await Box.async_from_host(api_host)
 
     async def test_unknown_product_without_device_section(self, aioclient_mock, data):
         host = "172.1.2.3"
@@ -43,13 +43,11 @@ class TestUnknownDevice:
             port = 80
             timeout = 2
             api_host = ApiHost(host, port, timeout, aioclient_mock, None, None)
-            await Products.async_from_host(api_host)
+            await Box.async_from_host(api_host)
 
     async def test_unknown_product_without_device_and_type(self, aioclient_mock, data):
         host = "172.1.2.3"
-        with pytest.raises(
-            error.UnsupportedBoxResponse, match=r"Missing 'type' field:"
-        ):
+        with pytest.raises(error.UnsupportedBoxResponse, match=r"has no type"):
             del data["type"]
             json_get_expect(
                 aioclient_mock, f"http://{host}:80/api/device/state", json=data
@@ -58,4 +56,4 @@ class TestUnknownDevice:
             port = 80
             timeout = 2
             api_host = ApiHost(host, port, timeout, aioclient_mock, None, None)
-            await Products.async_from_host(api_host)
+            await Box.async_from_host(api_host)